Adding word frequencies from QS 2012 notes
I was going through my notes, so I decided to run a quick word count
analysis. The frequencies are computed on a case-insensitive basis. Some
of the highlights:

- I'm not surprised to see Data (85) and Tracking (35) in the top 20 ;)
- What (45) wins out slightly over How (42), but why (7) and when (7)
  are tied.
- Self (22) wins out slightly over people (17).
- Many of the really interesting themes are in the mid-frequency range,
  below all the stop words: Habit (12), social (9), Effect (9), Behavior
  (9), sensors (8), cue (8), Us (8), Personal (8), Our (8), Design (8),
  stress (7), sleep (7), routine (7), process (7), Power (7), Platform
  (7), Community (7), control (6), Willpower (6), Wearable (6), Visual
  (6), subjective (5), motivation (5), location (5), Mindfulness (5).

I didn't run any synonym clustering, so the actual "concept frequencies"
would be different. That said, this was a fun (if somewhat
coarse-grained) way for me to sum up the conference in a UNIX
one-liner!